drop all convenience features of next-adapter in favor of manual setup
