--- packages/next-adapter/customize/index.d.ts
@@ -1,1 +1,0 @@
-export * from '../build/customize';
