--- packages/next-adapter/customize/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../build/customize');
