--- packages/next-adapter/template/default-gitignore
@@ -1,22 +1,0 @@
-node_modules/**/*
-.expo/*
-npm-debug.*
-*.jks
-*.p8
-*.p12
-*.key
-*.mobileprovision
-*.orig.*
-web-build/
-web-report/
-
-# debug
-yarn-debug.log*
-yarn-error.log*
-
-# macOS
-.DS_Store
-
-# misc
-.DS_Store
-.env*
