--- packages/next-adapter/template/next.config.js
@@ -1,7 +1,0 @@
-// Learn more: https://github.com/expo/expo/blob/master/docs/pages/versions/unversioned/guides/using-nextjs.md#withexpo
-
-const { withExpo } = require('@expo/next-adapter');
-
-module.exports = withExpo({
-  projectRoot: __dirname,
-});
