--- packages/next-adapter/template/pages/_document.js
@@ -1,1 +1,0 @@
-export { default } from '@expo/next-adapter/document';
